Instructions classes moved to dedicated folder
