+++ src/instructions/Else.h
@@ -1,7 +1,7 @@
 #ifndef ELSE_H
 #define ELSE_H
 
-#include "Node.h"
+#include "../modules/Node.h"
 
 /**
  * @brief Node of the tree which represent a condition if
