+++ src/instructions/Repeat.h
@@ -2,8 +2,8 @@
 #define REPEAT_H
 
 //special files no to forget to import
-#include "Node.h"
-#include "ConditionalExpression.h"
+#include "../modules/Node.h"
+#include "../modules/ConditionalExpression.h"
 
 /**
  * @brief Represent a node of the tree whcih will traduct the loop repeat (do ... while in c++)
